F01: Add implementation of PartenaireService
